+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Tests for `RocketResponse`.
 - `NetworkError` enum with error descriptions.
 - Tests for `RocketConfigurationResponse`.
+- `NetworkService` protocol.
+- Mocks fo testing `NetworkService`.
 
 ### Changed